Fixes folder refactor. [Fixes #108621210]

+++ Archivist/src/main/java/com/zorroa/archivist/service/FolderServiceImpl.java
@@ -1,14 +1,16 @@
 package com.zorroa.archivist.service;
 
+import com.google.common.collect.Lists;
+import com.zorroa.archivist.SecurityUtils;
 import com.zorroa.archivist.repository.FolderDao;
 import com.zorroa.archivist.sdk.domain.DuplicateElementException;
 import com.zorroa.archivist.sdk.domain.Folder;
 import com.zorroa.archivist.sdk.domain.FolderBuilder;
 import com.zorroa.archivist.sdk.service.FolderService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.access.AccessDeniedException;
 import org.springframework.stereotype.Component;
 
-import java.util.ArrayList;
 import java.util.List;
 
 @Component
@@ -23,8 +25,13 @@ public class FolderServiceImpl implements FolderService {
     }
 
     @Override
-    public List<Folder> getAll(int userId) {
-        return folderDao.getAll(userId);
+    public List<Folder> getAll() {
+        return folderDao.getAll();
+    }
+
+    @Override
+    public List<Folder> getAllShared() {
+        return folderDao.getAllShared();
     }
 
     @Override
@@ -34,9 +41,14 @@ public class FolderServiceImpl implements FolderService {
 
     @Override
     public List<Folder> getAllDecendents(Folder folder) {
-        ArrayList<Folder> decendents = new ArrayList<>();
         List<Folder> children = getChildren(folder);
+        if (children.isEmpty()) {
+            return Lists.newArrayListWithCapacity(0);
+        }
+
+        List<Folder> decendents = Lists.newArrayList();
         decendents.addAll(children);
+
         for (Folder child : children) {
             List<Folder> grandchildren = getAllDecendents(child);
             decendents.addAll(grandchildren);
@@ -46,6 +58,14 @@ public class FolderServiceImpl implements FolderService {
 
     @Override
     public synchronized Folder create(FolderBuilder builder) {
+
+        if (!builder.getParentId().equals(Folder.ROOT_ID)) {
+            Folder parent = folderDao.get(builder.getParentId());
+            if (parent.getUserId() != SecurityUtils.getUser().getId()) {
+                throw new AccessDeniedException("Invalid folder owner");
+            }
+        }
+
         if (folderDao.exists(builder.getParentId(), builder.getName())) {
             throw new DuplicateElementException(String.format("The folder '%s' already exists.", builder.getName()));
         }
@@ -54,12 +74,19 @@ public class FolderServiceImpl implements FolderService {
 
     @Override
     public boolean update(Folder folder, FolderBuilder builder) {
-        // Convert any names in the builder to ids?
+        if (folder.getUserId() != SecurityUtils.getUser().getId()) {
+            throw new AccessDeniedException("Invalid folder owner");
+        }
+
         return folderDao.update(folder, builder);
     }
 
     @Override
     public boolean delete(Folder folder) {
+        if (folder.getUserId() != SecurityUtils.getUser().getId()) {
+            throw new AccessDeniedException("Invalid folder owner");
+        }
+
         return folderDao.delete(folder);
     }
 }
